requirements/test.txt
# Test dependencies go here.
-r base.txt

flake8==3.8.4
django-test-plus==1.4.0
factory-boy==3.2.0
coverage==5.4
django-coverage-plugin==1.8.0

# pytest
pytest-django==4.1.0
pytest-sugar==0.9.4

# Selenium for UI testing
selenium==3.141.0

# Tblib for tracebacks
tblib==1.7.0

# BeautifulSoup for HTML testing
beautifulsoup4==4.9.3

# Coverage through Codacy
codacy-coverage==1.3.11

# Black for formatting
black==20.8b1

# Mock library for urllib3
urllib3-mock==0.3.3
